Change roster to component
