+++ client/src/actions/index.js
@@ -33,26 +33,18 @@ const updateEmployeeAvailability = (employee, newAvailabilities) => {
     return { user_id: employee.id, day_part_id: dayPartId, is_available: newAvailabilities[dayPartId] };
   });
 
-  const response = axios.post('/employee_availability', {
+  const response = axios.patch('/employee_availability', {
     employeeAvailabilities: requestBody,
-  }); // this will return a promise, which the redux-promise middleware will WAIT for before it sends the action to reducers
-  console.log('RESPONSE', response);
+  });
+
   return {
     type: 'UPDATE_EMPLOYEE_AVAILABILITY',
     payload: response,
   };
 };
 
-const selectEmployee = (employee) => {
-  return {
-    type: 'SELECT_EMPLOYEE',
-    payload: employee,
-  };
-};
-
 module.exports = {
   updateEmployeeAvailability: updateEmployeeAvailability,
-  selectEmployee: selectEmployee,
   getAllUsers: getAllUsers,
   getAllEmployeeAvailabilities: getAllEmployeeAvailabilities,
   getAllDayParts: getAllDayParts,
